backend/pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["Ayush Bhattarai"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.2"
youtube-transcript-api = "^0.6.3"
groq = "^0.4.2"
yt-dlp = "^2024.3.10"
openai = "^1.54.4"
ffmpeg = "^1.4"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
